backend_core/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
torch>=2.1.0
transformers>=4.36.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
django-cors-headers>=4.0.0
